% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appendGIFs.R
\name{appendGIFs}
\alias{appendGIFs}
\title{appendGIFs read and append two GIFs}
\usage{
appendGIFs(gif_1, gif_2, vertical = FALSE)
}
\arguments{
\item{gif_1}{First GIF image}

\item{gif_2}{Second GIF image}

\item{vertical}{TRUE: GIFs top-to-bottom, FALSE: left-to-right (default)}
}
\description{
appendGIFs read and append two GIFs
}
\keyword{GIFs}
\keyword{append}
