% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areaQuadrantTime.R
\name{areaQuadrantTime}
\alias{areaQuadrantTime}
\title{Area chart for percentage of quadrant time}
\usage{
areaQuadrantTime(
  data,
  id,
  day,
  trial,
  centerx,
  centery,
  radius = 75,
  platformx,
  platformy,
  platformradius = 7.5,
  device = "jpeg",
  width = 18,
  height = 12,
  units = "cm",
  dpi = 300,
  viridis_color_palette = "D",
  theme_settings = NULL
)
}
\arguments{
\item{data}{Data set containing at least following columns: "Time", "x", "y", "Animal", "Day", "Trial".
"x" and "y" represent the coordinates (position) of the animal at a certain timepoint ("Time") during the trial.}

\item{id}{ID of the animal}

\item{day}{day number}

\item{trial}{trial number}

\item{centerx}{x coordinate of the center of the morris water maze (cm)}

\item{centery}{y coordinate of the center of the morris water maze (cl)}

\item{radius}{radius of the morris water maze (cm), default = 75}

\item{platformx}{x coordinate of the center of the platform (cm)}

\item{platformy}{y coordinate of the center of the platform (cm)}

\item{platformradius}{radius of the platform (cm), default = 7.5}

\item{device}{Device for the graph. Options: "png", "eps", "ps", "tex", "pdf", "jpeg", "tiff", "png", "bmp", "svg", "wmf"}

\item{width}{Width of the graph, default = 18}

\item{height}{Height of the graph, default = 12}

\item{units}{Units. Options: "in", "cm", "mm". Default = "cm".}

\item{dpi}{Resolution. Default = 300}

\item{viridis_color_palette}{Viridus color palette. Options: "A", "B", "C", "D", "E". Default = "D".}

\item{theme_settings}{Optional parameter that passes list of arguments to ggplot2's theme() function.}
}
\description{
areaQuadrantTime creates a geom_area plot showing the percentage of time spent per quadrant for a specific trial of one animal, visualized in a stacked area plot.
This function returns the graph as a ggplot object and saves the image (default = jpeg).
}
\keyword{area}
\keyword{graph}
\keyword{percentage}
\keyword{quadrant}
\keyword{stacked}
\keyword{time}
