% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barGIF.R
\name{barGIF}
\alias{barGIF}
\title{barGIF creates a GIF showing the percentage of time spent per quadrant for one trial of one animal}
\usage{
barGIF(
  data,
  id,
  day,
  trial,
  centerx,
  centery,
  radius,
  platformx,
  platformy,
  platformradius,
  removeSwimspeedOutliers = TRUE,
  SwimspeedLimit = 50,
  loop = FALSE,
  width = 480,
  height = 480,
  fps = 10,
  duration = 10,
  frames = 100,
  time_bins = 50
)
}
\arguments{
\item{data}{Data set containing at least following columns: "Time", "x", "y", "Animal", "Day", "Trial".
"x" and "y" represent the coordinates (position) of the animal at a certain timepoint ("Time") during the trial.}

\item{id}{ID of the animal}

\item{day}{day number}

\item{trial}{trial number}

\item{centerx}{x coordinate of the center of the morris water maze}

\item{centery}{y coordinate of the center of the morris water maze}

\item{radius}{radius of the morris water maze}

\item{platformx}{x coordinate of the center of the platform}

\item{platformy}{y coordinate of the center of the platform}

\item{platformradius}{radius of the platform}

\item{removeSwimspeedOutliers}{removes swim speed outliers (>= SwimspeedLimit), default = TRUE}

\item{SwimspeedLimit}{default = 50 (cm/s), only used when removeSwimspeedOutliers = TRUE}

\item{loop}{Loop the animation, default = FALSE}

\item{width}{Width of the animation in pixels, default = 480}

\item{height}{Height of the animation in pixels, default = 480}

\item{fps}{Frames per second of the animation, default = 10}

\item{duration}{Duration of the animation in seconds, default = 10}

\item{frames}{Number of frames in the animation, default = 100}

\item{time_bins}{Number of time-bins in the animation, default = 50}
}
\description{
barGIF creates a GIF showing the percentage of time spent per quadrant for one trial of one animal
}
\keyword{bar}
\keyword{graph}
\keyword{quadrant}
\keyword{time}
