% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circle.R
\name{circle}
\alias{circle}
\title{Create data set for plotting circles}
\usage{
circle(
  x = 0,
  y = 0,
  radius = 75,
  nrow_data = 100,
  from = 0,
  to = 2,
  add_center = FALSE
)
}
\arguments{
\item{x}{The x coordinate of the center of the circle.}

\item{y}{The y coordinate of the center of the circle.}

\item{radius}{The radius of the circle.}

\item{nrow_data}{The number of data points.}

\item{from}{Start of sequence. Default = 0}

\item{to}{End of sequence. E.g. End = 0.5 results in a quadrant polygon. Default = 2}

\item{add_center}{Add the coordinate of the center (x,y) to the data set. Useful when creating quadrant polygons. Default = FALSE}
}
\description{
The circle function creates a data set with x and y values that can be used as coordinates to plot a circle.
The arguments 'from' and 'to' can be adjust to output different quadrants of the circle.
}
