% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupCOORD.R
\name{groupCOORD}
\alias{groupCOORD}
\title{Group coordinates data}
\usage{
groupCOORD(data, by)
}
\arguments{
\item{data}{Data set containing at least following columns: "Time", "x", "y", "Animal", "Day", "Trial".
"x" and "y" represent the coordinates (position) of the animal at a certain timepoint ("Time") during the trial.}

\item{by}{Variable to group by. E.g. "Trial", c("Group", "Trial").}
}
\description{
The groupCOORD function uses dplyr to group data over groups, trials and/or days. Note that data will always be
grouped over time, which might only make sense for trials of equal length (e.g. probe trials).
}
