% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_data-data.R
\docType{data}
\name{track_data}
\alias{track_data}
\title{Track data of 4 trials in a Morris water maze experiment}
\format{A data frame with 81889 rows and 7 columns.
\describe{
\itemize{
  \item Time: Timestamp
  \item x: x coordinate in Morris water maze setup
  \item y: y coordinate in Morris water maze setup
  \item Animal: Animal ID
  \item Day: Day of experiment (numeric)
  \item Trial: Trial number
  \item Group: Group information (NA)
}
}}
\usage{
data(track_data)
}
\description{
Track data of 4 trials in a Morris water maze experiment
}
\examples{
data(track_data)
}
\keyword{Morris}
\keyword{datasets}
\keyword{maze}
\keyword{mwm}
\keyword{track}
\keyword{water}
