% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackGIF.R
\name{trackGIF}
\alias{trackGIF}
\title{trackGIF creates a gif showing the track, distance to target and velocity of one specific trial}
\usage{
trackGIF(data, id, day, trial, centerx, centery, radius, platformx,
  platformy, platformradius, removeSwimspeedOutliers = TRUE,
  SwimspeedLimit = 50)
}
\arguments{
\item{data}{data set containing at least following columns: "Time", "x", "y", "Animal", "Day", "Trial"}

\item{id}{ID of the animal}

\item{day}{day number}

\item{trial}{trial number}

\item{centerx}{x coordinate of the center of the morris water maze}

\item{centery}{y coordinate of the center of the morris water maze}

\item{radius}{radius of the morris water maze}

\item{platformx}{x coordinate of the center of the platform}

\item{platformy}{y coordinate of the center of the platform}

\item{platformradius}{radius of the platform}

\item{removeSwimspeedOutliers}{removes swim speed outliers (>= SwimspeedLimit), default = TRUE}

\item{SwimspeedLimit}{default = 50 (cm/s), only used when removeSwimspeedOutliers = TRUE}
}
\description{
Output: 4 gifs (track, distance to target, velocity and all three combined)
}
\keyword{distance}
\keyword{gif}
\keyword{target}
\keyword{to}
\keyword{track}
\keyword{velocity}
