% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateCOORD.R
\name{updateCOORD}
\alias{updateCOORD}
\title{Preprocess coordinates data}
\usage{
updateCOORD(
  data,
  centerx,
  centery,
  radius = 75,
  platformx,
  platformy,
  platformradius = 7.5,
  removeNA = TRUE
)
}
\arguments{
\item{data}{Data set containing at following columns: "x", "y".
"x" and "y" represent the coordinates (position) of the animal at a certain timepoint during a trial.}

\item{centerx}{x coordinate of the center of the morris water maze (cm).}

\item{centery}{y coordinate of the center of the morris water maze (cl).}

\item{radius}{radius of the morris water maze (cm), default = 75.}

\item{platformx}{x coordinate of the center of the platform (cm).}

\item{platformy}{y coordinate of the center of the platform (cm).}

\item{platformradius}{radius of the platform (cm), default = 7.5.}

\item{removeNA}{Remove NAs. Default=TRUE.}
}
\description{
The updateCOORD function rescales x and y coordinates to center (0,0) and adds quadrant information to the data set.
Adjacent Left is the quadrant to the left of the target quadrant when standing at the target quadrant and looking towards the platform.
Adjacent Right is the quadrant to the right of the target quadrant when standing at the target quadrant and looking towards the platform.
}
