% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barGIF.R
\name{barGIF}
\alias{barGIF}
\title{Bar chart for percentage of quadrant time}
\usage{
barGIF(
  data,
  id,
  day,
  trial,
  centerx,
  centery,
  radius = 75,
  platformx,
  platformy,
  platformradius = 7.5,
  bar_colours = c("#F8766D", "#7CAE00", "#00BFC4", "#C77CFF"),
  show_time = FALSE,
  loop = FALSE,
  width = 480,
  height = 480,
  fps = 10,
  duration = 10,
  frames = 100,
  time_bins = 50,
  theme_settings = NULL,
  title = NA
)
}
\arguments{
\item{data}{Data set containing at least following columns: "Time", "x", "y", "Animal", "Day", "Trial".
"x" and "y" represent the coordinates (position) of the animal at a certain timepoint ("Time") during the trial.}

\item{id}{ID of the animal}

\item{day}{day number}

\item{trial}{trial number}

\item{centerx}{x coordinate of the center of the morris water maze (cm)}

\item{centery}{y coordinate of the center of the morris water maze (cl)}

\item{radius}{radius of the morris water maze (cm), default = 75}

\item{platformx}{x coordinate of the center of the platform (cm)}

\item{platformy}{y coordinate of the center of the platform (cm)}

\item{platformradius}{radius of the platform (cm), default = 7.5}

\item{bar_colours}{Colours of the four bars. Default = c("#F8766D","#7CAE00", "#00BFC4", "#C77CFF")}

\item{show_time}{Show time-bins. Default = FALSE}

\item{loop}{Loop the animation, default = FALSE}

\item{width}{Width of the animation (px), default = 480}

\item{height}{Height of the animation (px), default = 480}

\item{fps}{Frames per second of the animation, default = 10}

\item{duration}{Duration of the animation(s), default = 10}

\item{frames}{Number of frames in the animation, default = 100}

\item{time_bins}{Number of time-bins in the animation, default = 50}

\item{theme_settings}{Optional parameter that passes list of arguments to ggplot2's theme() function.}

\item{title}{Add title to GIF. Default = NA}
}
\description{
The barGIF function creates a GIF showing the percentage of time spent per quadrant for a specific trial of one animal.
}
\keyword{bar}
\keyword{graph}
\keyword{quadrant}
\keyword{time}
