% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapGIF.R
\name{heatmapGIF}
\alias{heatmapGIF}
\title{Morris water maze track data GIF}
\usage{
heatmapGIF(
  data,
  id,
  day,
  trial,
  centerx,
  centery,
  radius = 75,
  platformx,
  platformy,
  platformradius = 7.5,
  ndata_circle = 100,
  remove_data_outside_maze = TRUE,
  platform_colour = NA,
  platform_alpha = 1,
  platform_linetype = "dotted",
  platform_line_colour = "black",
  platform_line_size = 1,
  heatmap_low = "yellow",
  heatmap_high = "red",
  type = "raster",
  interpolate = TRUE,
  contour_filled = TRUE,
  contour_colour_scaled = FALSE,
  contour_colour_filled = NA,
  contour_colour = "blue",
  loop = FALSE,
  width = 480,
  height = 480,
  duration = 10,
  frames = 100,
  resolution = 80,
  theme_settings = NULL
)
}
\arguments{
\item{data}{Data set containing at least following columns: "Time", "x", "y", "Animal", "Day", "Trial".
"x" and "y" represent the coordinates (position) of the animal at a certain timepoint ("Time") during the trial.}

\item{id}{ID of the animal}

\item{day}{day number}

\item{trial}{trial number}

\item{centerx}{x coordinate of the center of the morris water maze (cm)}

\item{centery}{y coordinate of the center of the morris water maze (cl)}

\item{radius}{radius of the morris water maze (cm), default = 75}

\item{platformx}{x coordinate of the center of the platform (cm)}

\item{platformy}{y coordinate of the center of the platform (cm)}

\item{platformradius}{radius of the platform (cm), default = 7.5}

\item{ndata_circle}{Number of data points in the circle data set. Higher means smoother (more perfect) circle. Default = 100}

\item{remove_data_outside_maze}{Remove datapoints that lie outside the water maze. Default = TRUE}

\item{platform_colour}{Colour of the platform. Name or hexadecimal code (e.g.: #FF1020). Default = NA}

\item{platform_alpha}{Alpha level for platform. Default = 1}

\item{platform_linetype}{Linetype for platform. Derived from ggplot2. Default = "dotted"}

\item{platform_line_colour}{Colour of platform circle. Default = "black"}

\item{platform_line_size}{Size of platform line. Default = 1}

\item{heatmap_low}{Low range colour heatmap. Default = "yellow"}

\item{heatmap_high}{High range colour heatmap. Default = "red"}

\item{type}{Type of heatmap. Options are "raster" or "contour". Default = "raster"}

\item{interpolate}{Interpolate raster heatmaps? Ignored for contour heatmaps. Default = TRUE}

\item{contour_filled}{Fill contour heatmaps? Ignored for raster heatmaps. Default = TRUE}

\item{contour_colour_scaled}{Colour scale for contour heatmaps. When false, colour of contour lines is set by contour_colour.
When TRUE, colour scale is set by heatmap_low and heatmap_high. Default = FALSE}

\item{contour_colour_filled}{Colour of contour lines in filled contour heatmap. Default = NA}

\item{contour_colour}{Colour of contour lines in empty contour heatmap. Default = "blue"}

\item{loop}{Loop the animation, default = FALSE}

\item{width}{Width of the animation (px), default = 480}

\item{height}{Height of the animation (px), default = 480}

\item{duration}{Duration of the animation(s), default = 10}

\item{frames}{Number of frames in the animation, default = 100}

\item{resolution}{Resolution of GIF, passed to gifski. Default = 80}

\item{theme_settings}{Optional parameter that passes list of arguments to ggplot2's theme() function.}
}
\description{
The trackGIF function creates a GIF showing the track of the animal for a specific trial.
}
\keyword{distance}
\keyword{gif}
\keyword{target}
\keyword{to}
\keyword{track}
\keyword{velocity}
