% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackGIF.R
\name{trackGIF}
\alias{trackGIF}
\title{trackGIF creates a GIF showing the track of the animal for a specific trial}
\usage{
trackGIF(
  data,
  id,
  day,
  trial,
  centerx,
  centery,
  radius = 75,
  platformx,
  platformy,
  platformradius = 7.5,
  ndata_circle = 100,
  quadrant_colours = c("white", "white", "white", "white"),
  platform_colour = "grey",
  alpha_quadrants = 0.2,
  alpha_platform = 1,
  track_colour = "orange",
  alpha_track = 0.35,
  loop = FALSE,
  width = 480,
  height = 480,
  fps = 10,
  duration = 10,
  frames = 100,
  time_bins = 50
)
}
\arguments{
\item{data}{Data set containing at least following columns: "Time", "x", "y", "Animal", "Day", "Trial".
"x" and "y" represent the coordinates (position) of the animal at a certain timepoint ("Time") during the trial.}

\item{id}{ID of the animal}

\item{day}{day number}

\item{trial}{trial number}

\item{centerx}{x coordinate of the center of the morris water maze (cm)}

\item{centery}{y coordinate of the center of the morris water maze (cl)}

\item{radius}{radius of the morris water maze (cm), default = 75}

\item{platformx}{x coordinate of the center of the platform (cm)}

\item{platformy}{y coordinate of the center of the platform (cm)}

\item{platformradius}{radius of the platform (cm), default = 7.5}

\item{ndata_circle}{Number of data points in the circle data set. Higher means smoother (more perfect) circle. Default = 100}

\item{quadrant_colours}{Fill colours of quadrants. Order = top left, top right, bottom left, bottom right. Name or hexadecimal code (e.g.: #FF1020). Default = c("white","white","white","white")}

\item{platform_colour}{Colour of the platform. Name or hexadecimal code (e.g.: #FF1020). Default = "grey"}

\item{alpha_quadrants}{Alpha level for quadrants. Default = 0.2}

\item{alpha_platform}{Alpha level for platform. Default = 1}

\item{track_colour}{Colour of the track line. Name or hexadecimal code (e.g.: #FF1020). Default = orange}

\item{alpha_track}{Alpha level for the track line. Default = 0.35}

\item{loop}{Loop the animation, default = FALSE}

\item{width}{Width of the animation (px), default = 480}

\item{height}{Height of the animation (px), default = 480}

\item{fps}{Frames per second of the animation, default = 10}

\item{duration}{Duration of the animation(s), default = 10}

\item{frames}{Number of frames in the animation, default = 100}

\item{time_bins}{Number of time-bins in the animation, default = 50}
}
\description{
trackGIF creates a GIF showing the track of the animal for a specific trial
}
\keyword{distance}
\keyword{gif}
\keyword{target}
\keyword{to}
\keyword{track}
\keyword{velocity}
