% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackGIF.R
\name{trackGIF}
\alias{trackGIF}
\title{trackGIF creates 3 GIFs: track, distance to target and velocity}
\usage{
trackGIF(
  data,
  id,
  day,
  trial,
  centerx,
  centery,
  radius,
  platformx,
  platformy,
  platformradius,
  removeSwimspeedOutliers = TRUE,
  SwimspeedLimit = 50
)
}
\arguments{
\item{data}{Data set containing at least following columns: "Time", "x", "y", "Animal", "Day", "Trial".
"x" and "y" represent the coordinates (position) of the animal at a certain timepoint ("Time") during the trial.}

\item{id}{ID of the animal}

\item{day}{day number}

\item{trial}{trial number}

\item{centerx}{x coordinate of the center of the morris water maze}

\item{centery}{y coordinate of the center of the morris water maze}

\item{radius}{radius of the morris water maze}

\item{platformx}{x coordinate of the center of the platform}

\item{platformy}{y coordinate of the center of the platform}

\item{platformradius}{radius of the platform}

\item{removeSwimspeedOutliers}{removes swim speed outliers (>= SwimspeedLimit), default = TRUE}

\item{SwimspeedLimit}{default = 50 (cm/s), only used when removeSwimspeedOutliers = TRUE}
}
\description{
trackGIF creates 3 GIFs: track, distance to target and velocity
}
\keyword{distance}
\keyword{gif}
\keyword{target}
\keyword{to}
\keyword{track}
\keyword{velocity}
