% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackGIF.R
\name{trackGIF}
\alias{trackGIF}
\title{Morris water maze track data GIF}
\usage{
trackGIF(
  data,
  id,
  day,
  trial,
  centerx,
  centery,
  radius = 75,
  platformx,
  platformy,
  platformradius = 7.5,
  ndata_circle = 100,
  quadrant_colours = c("white", "white", "white", "white"),
  quadrants_alpha = 0.2,
  platform_alpha = 1,
  platform_colour = "black",
  platform_line_size = 0.5,
  platform_linetype = "solid",
  platform_line_colour = "black",
  track_colour = "orange",
  track_alpha = 0.35,
  loop = FALSE,
  width = 480,
  height = 480,
  fps = 10,
  duration = 10,
  theme_settings = NULL,
  title = NA,
  plot_original_platform = FALSE,
  original_platformx = NULL,
  original_platformy = NULL,
  original_platform_colour = "grey",
  original_platform_alpha = 0.4,
  original_platform_linetype = "dotted",
  original_platform_line_size = 0.5,
  original_platform_line_colour = "black"
)
}
\arguments{
\item{data}{Data set containing at least following columns: "Time", "x", "y", "Animal", "Day", "Trial".
"x" and "y" represent the coordinates (position) of the animal at a certain timepoint ("Time") during the trial.}

\item{id}{ID of the animal}

\item{day}{day number}

\item{trial}{trial number}

\item{centerx}{x coordinate of the center of the morris water maze (cm)}

\item{centery}{y coordinate of the center of the morris water maze (cl)}

\item{radius}{radius of the morris water maze (cm), default = 75}

\item{platformx}{x coordinate of the center of the platform (cm)}

\item{platformy}{y coordinate of the center of the platform (cm)}

\item{platformradius}{radius of the platform (cm), default = 7.5}

\item{ndata_circle}{Number of data points in the circle data set. Higher means smoother (more perfect) circle. Default = 100}

\item{quadrant_colours}{Fill colours of quadrants. Order = top left, top right, bottom left, bottom right. Name or hexadecimal code (e.g.: #FF1020). Default = c("white","white","white","white")}

\item{quadrants_alpha}{Alpha level for quadrants. Default = 0.2}

\item{platform_alpha}{Alpha level for platform. Default = 1}

\item{platform_colour}{Colour of the platform. Name or hexadecimal code (e.g.: #FF1020). Default = "black"}

\item{platform_line_size}{Size of platform circle. Default = 0.5}

\item{platform_linetype}{Linetype of platform circle. Default = "solid"}

\item{platform_line_colour}{Colour of platform circle line. Default = "black"}

\item{track_colour}{Colour of the track line. Name or hexadecimal code (e.g.: #FF1020). Default = orange}

\item{track_alpha}{Alpha level for the track line. Default = 0.35}

\item{loop}{Loop the animation, default = FALSE}

\item{width}{Width of the animation (px), default = 480}

\item{height}{Height of the animation (px), default = 480}

\item{fps}{Frames per second of the animation. default = 10}

\item{duration}{Duration of the animation(s), default = 10}

\item{theme_settings}{Optional parameter that passes list of arguments to ggplot2's theme() function.}

\item{title}{Add title to GIF. Default = NA}

\item{plot_original_platform}{Plot the original platform (for reversal trials). Default = FALSE}

\item{original_platformx}{x coordinate of the center of the original platform (cm). Ignored if plot_original_platform = FALSE}

\item{original_platformy}{y coordinate of the center of the original platform (cm). Ignored if plot_original_platform = FALSE}

\item{original_platform_colour}{Colour of the original platform. Name or hexadecimal code (e.g.: #FF1020). Ignored if plot_original_platform = FALSE. Default = "grey"}

\item{original_platform_alpha}{Alpha level for original platform. Ignored if plot_original_platform = FALSE. Default = 0.4}

\item{original_platform_linetype}{Linetype of original platform circle. Ignored if plot_original_platform = FALSE. Default = "dotted"}

\item{original_platform_line_size}{Size of original platform circle. Ignored if plot_original_platform = FALSE. Default = 0.5}

\item{original_platform_line_colour}{Colour of original platform circle line. Ignored if plot_original_platform = FALSE. Default = "black"}
}
\description{
The trackGIF function creates a GIF showing the track of the animal for a specific trial.
}
\keyword{gif}
\keyword{maze}
\keyword{morris}
\keyword{reversal}
\keyword{track}
\keyword{water}
